pyproject.toml
[tool.poetry]
name = "talmin"
version = "0.1.0"
description = ""
authors = ["Christoph Vos <christoph.a.vos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.15.0"
dash-ag-grid = "^31.0.1"
dash-bootstrap-components = "^1.5.0"
ipykernel = "^6.29.2"
ipywidgets = "^8.1.2"
networkx = "^3.2.1"
numba = "^0.59.0"
numpy = "^1.26.4"
pandas = "^2.2.1"
polars = "^0.20.10"
pyarrow = "^15.0.0"
scikit-learn = "^1.4.1.post1"
tqdm = "^4.66.2"
umap-learn = "^0.5.5"
xgboost = "^2.0.3"
dash-bootstrap-templates = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
